% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcppcuda.R
\name{hello}
\alias{hello}
\title{Function \code{hello}}
\usage{
hello()
}
\description{
Main function of the package. Passes a \code{MyClass}
object from R to C++, runs som CUDA code, changes, the object,
and prints the result
}

