% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcppcuda.R
\docType{class}
\name{MyClass-class}
\alias{MyClass}
\alias{MyClass-class}
\title{Class \code{MyClass}}
\description{
A toy class for figuring out how to pass s4 objects from R to C++.
}
\section{Slots}{

\describe{
\item{\code{x}}{a numeric vector}

\item{\code{y}}{an integer vector}
}}

