% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcppcuda.R
\name{rsetDevice}
\alias{rsetDevice}
\title{Function \code{rsetDevice}}
\usage{
rsetDevice(device)
}
\arguments{
\item{device}{Integer index of a CUDA-capable GPU. Must be >= 0 and < number of GPUs.}
}
\description{
Choose a CUDA-capable GPU to use for the MCMC.
}

