% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{someCPPcode}
\alias{someCPPcode}
\title{Function \code{someCPPcode}}
\usage{
someCPPcode(r)
}
\arguments{
\item{r}{a \code{MyClass} object.}
}
\description{
Run some C++ code to change
the slots in argument object \code{r}.
}

